requirements.txt
Flask
mysql-connector
names
